+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/HubdleKotlinAndroidExtension.kt
@@ -2,8 +2,8 @@ package com.javiersc.hubdle.extensions.kotlin.android
 
 import com.javiersc.hubdle.extensions.HubdleDslMarker
 import com.javiersc.hubdle.extensions._internal.PluginIds
-import com.javiersc.hubdle.extensions.kotlin.android.application.KotlinAndroidApplicationExtension
-import com.javiersc.hubdle.extensions.kotlin.android.library.KotlinAndroidLibraryExtension
+import com.javiersc.hubdle.extensions.kotlin.android.application.HubdleKotlinAndroidApplicationExtension
+import com.javiersc.hubdle.extensions.kotlin.android.library.HubdleKotlinAndroidLibraryExtension
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
@@ -11,13 +11,13 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
 
 @HubdleDslMarker
-public open class KotlinAndroidExtension @Inject constructor(objects: ObjectFactory) {
+public open class HubdleKotlinAndroidExtension @Inject constructor(objects: ObjectFactory) {
 
-    public val application: KotlinAndroidApplicationExtension = objects.newInstance()
+    public val application: HubdleKotlinAndroidApplicationExtension = objects.newInstance()
 
     @HubdleDslMarker
     public fun Project.application(
-        action: Action<in KotlinAndroidApplicationExtension> = Action {}
+        action: Action<in HubdleKotlinAndroidApplicationExtension> = Action {}
     ) {
         pluginManager.apply(PluginIds.Android.application)
         pluginManager.apply(PluginIds.Android.kotlin)
@@ -25,10 +25,10 @@ public open class KotlinAndroidExtension @Inject constructor(objects: ObjectFact
         action.execute(application)
     }
 
-    public val library: KotlinAndroidLibraryExtension = objects.newInstance()
+    public val library: HubdleKotlinAndroidLibraryExtension = objects.newInstance()
 
     @HubdleDslMarker
-    public fun Project.library(action: Action<in KotlinAndroidLibraryExtension> = Action {}) {
+    public fun Project.library(action: Action<in HubdleKotlinAndroidLibraryExtension> = Action {}) {
         pluginManager.apply(PluginIds.Android.library)
         pluginManager.apply(PluginIds.Android.kotlin)
         library.run { isEnabled = true }