+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/intellij/HubdleIntellijExtension.kt
@@ -1,12 +1,13 @@
-package com.javiersc.hubdle.extensions.kotlin.jvm
+package com.javiersc.hubdle.extensions.kotlin.intellij
 
 import com.javiersc.hubdle.extensions.HubdleDslMarker
-import com.javiersc.hubdle.extensions._internal.PluginIds
 import com.javiersc.hubdle.extensions._internal.state.hubdleState
 import com.javiersc.hubdle.extensions.kotlin.MainAndTestKotlinSourceSetsOptions
+import com.javiersc.hubdle.extensions.kotlin.jvm.KotlinJvmOptions
 import com.javiersc.hubdle.extensions.kotlin.jvm._internal.jvmFeatures
 import com.javiersc.hubdle.extensions.options.EnableableOptions
 import com.javiersc.hubdle.extensions.options.FeaturesOptions
+import com.javiersc.hubdle.extensions.options.GradleDependenciesOptions
 import com.javiersc.hubdle.extensions.options.RawConfigOptions
 import com.javiersc.hubdle.extensions.options.SourceDirectoriesOptions
 import javax.inject.Inject
@@ -14,42 +15,60 @@ import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Project
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.plugins.JavaApplication
 import org.gradle.kotlin.dsl.newInstance
 import org.gradle.kotlin.dsl.the
+import org.jetbrains.intellij.IntelliJPluginExtension
+import org.jetbrains.intellij.tasks.PatchPluginXmlTask
+import org.jetbrains.intellij.tasks.PublishPluginTask
+import org.jetbrains.intellij.tasks.SignPluginTask
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 @HubdleDslMarker
-public open class KotlinJvmExtension
+public open class HubdleIntellijExtension
 @Inject
 constructor(
     objects: ObjectFactory,
 ) :
     EnableableOptions,
-    FeaturesOptions<KotlinJvmExtension.FeaturesExtension>,
+    FeaturesOptions<HubdleIntellijExtension.FeaturesExtension>,
     KotlinJvmOptions,
     SourceDirectoriesOptions<KotlinSourceSet>,
-    RawConfigOptions<KotlinJvmExtension.RawConfigExtension>,
-    MainAndTestKotlinSourceSetsOptions<KotlinSourceSet> {
+    RawConfigOptions<HubdleIntellijExtension.RawConfigExtension>,
+    MainAndTestKotlinSourceSetsOptions<KotlinSourceSet>,
+    GradleDependenciesOptions {
 
     override var Project.isEnabled: Boolean
-        get() = hubdleState.kotlin.jvm.isEnabled
-        set(value) = hubdleState.kotlin.jvm.run { isEnabled = value }
+        get() = hubdleState.kotlin.intellij.isEnabled
+        set(value) = hubdleState.kotlin.intellij.run { isEnabled = value }
 
     override val features: FeaturesExtension = objects.newInstance()
 
     @HubdleDslMarker
     override fun features(action: Action<FeaturesExtension>): Unit = super.features(action)
 
-    @HubdleDslMarker
-    public fun Project.application(action: Action<JavaApplication>) {
-        pluginManager.apply(PluginIds.Gradle.application)
-        hubdleState.kotlin.jvm.application = action
-    }
-
     override val Project.sourceSets: NamedDomainObjectContainer<KotlinSourceSet>
         get() = the<KotlinJvmProjectExtension>().sourceSets
+
+    @HubdleDslMarker
+    public fun Project.intellij(action: Action<IntelliJPluginExtension> = Action {}) {
+        hubdleState.kotlin.intellij.intellij = action
+    }
+
+    @HubdleDslMarker
+    public fun Project.patchPluginXml(action: Action<PatchPluginXmlTask> = Action {}) {
+        hubdleState.kotlin.intellij.patchPluginXml = action
+    }
+
+    @HubdleDslMarker
+    public fun Project.publishPlugin(action: Action<PublishPluginTask> = Action {}) {
+        hubdleState.kotlin.intellij.publishPlugin = action
+    }
+
+    @HubdleDslMarker
+    public fun Project.signPlugin(action: Action<SignPluginTask> = Action {}) {
+        hubdleState.kotlin.intellij.signPlugin = action
+    }
 
     @HubdleDslMarker
     override fun Project.main(action: Action<KotlinSourceSet>) {