+++ hubdle-gradle-plugin/main/kotlin/composeExtension.accessor.kt
@@ -1,0 +1,7 @@
+@file:Suppress("UnusedReceiverParameter")
+
+import org.jetbrains.compose.ComposePlugin
+import org.jetbrains.kotlin.gradle.plugin.KotlinDependencyHandler
+
+public val KotlinDependencyHandler.compose: ComposePlugin.Dependencies
+    get() = ComposePlugin.Dependencies